Refactor and update database configurations
- Explicitly cast `TotalPrice` in `ShoppingCart` to `decimal`.
- Cleaned up gRPC client configuration by removing commented code.
- Updated `DiscountContext` to comment out primary key setup for `Coupon`.
- Added `Microsoft.EntityFrameworkCore.Design` package reference.
- Introduced new connection string for `OrderDb` in `appsettings.json`.
- Removed unnecessary using directives in `ApplicationDbContext` and configuration classes.
- Changed database connection string in `DependencyInjection` to `OrderDb`.
- Added `orderdb` service to `docker-compose.override.yml` with SQL Server config.
- Specified SQL Server image for `orderdb` in `docker-compose.yml`.
- Created `GlobalUsing.cs` for global using directives to simplify code.

Branch: feature/update-database-config

+++ src/Services/Basket/Basket.API/Models/ShoppingCart.cs
@@ -8,7 +8,7 @@ namespace Basket.API.Models
 
         public List<ShoppingCartItem> Items { get; set; } = new();
 
-        public decimal TotalPrice => Items.Sum(x => x.Price * x.Quantity);
+        public decimal TotalPrice => (decimal)Items.Sum(x => x.Price * x.Quantity);
 
         public ShoppingCart(string userName)
         {
